增加测试案例并使用INI文件

+++ Arithmetic/A/add.h
@@ -1,1 +1,1 @@
-extern "C" int add(int added, int add);
+extern "C" long long add(int added, int add);
